requirements.txt
fastapi
uvicorn
pyjwt[crypto]
passlib[bcrypt]
pydantic
python-dotenv 
mysql-connector-python

# pip install -r requirements.txt
# Pour lancer le serveur :
# python -m uvicorn main:app --host 0.0.0.0 --port 8000 --reload